// src/interfaces/foreground.ts
export interface Foreground {
    readonly base: string;
    readonly divider: string;
    readonly dividers: string;
    readonly disabled: string;
    readonly 'disabled-button': string;
    readonly 'disabled-text': string;
    readonly 'hint-text': string;
    readonly 'secondary-text': string;
    readonly icon: string;
    readonly icons: string;
    readonly text: string;
    readonly 'slider-min': string;
    readonly 'slider-off': string;
    readonly 'slider-off-active': string;
}