// src/interfaces/background.ts
export interface Background {
    readonly 'status-bar': string;
    readonly 'app-bar': string;
    readonly background: string;
    readonly hover: string;
    readonly card: string;
    readonly dialog: string;
    readonly 'disabled-button': string;
    readonly 'raised-button': string;
    readonly 'focused-button': string;
    readonly 'selected-button': string;
    readonly 'selected-disabled-button': string;
    readonly 'disabled-button-toggle': string;
    readonly 'unselected-chip': string;
    readonly 'disabled-list-option': string;
}